chore: remove unused files
