--- hw4/Yojahuang/includes/TileHelper.h
@@ -1,15 +1,0 @@
-#ifndef TILEHELPER_H
-#define TILEHELPER_H
-
-#include <vector>
-#include "matrix.h"
-#include <cstddef>
-
-class TileHelper
-{
-public:
-    Matrix padMatrix(size_t, const Matrix&);
-private:
-}; /* end class Line */
-
-#endif
