+++ hw4/Yojahuang/matrix_mul.cpp
@@ -1,6 +1,3 @@
-#include "matrix.h"
-#include "TileHelper.h"
-
 #include <stdexcept>
 #include <cstddef>
 
@@ -13,6 +10,296 @@
 #include <mkl/mkl_lapacke.h>
 
 namespace py = pybind11;
+
+struct ByteCounterImpl
+{
+
+    std::atomic_size_t allocated = 0;
+    std::atomic_size_t deallocated = 0;
+    std::atomic_size_t refcount = 0;
+
+}; /* end struct ByteCounterImpl */
+
+/**
+ * One instance of this counter is shared among a set of allocators.
+ *
+ * The counter keeps track of the bytes allocated and deallocated, and report
+ * those two numbers in addition to bytes that remain allocated.
+ */
+class ByteCounter
+{
+
+public:
+
+    ByteCounter()
+      : m_impl(new ByteCounterImpl)
+    { incref(); }
+
+    ByteCounter(ByteCounter const & other)
+      : m_impl(other.m_impl)
+    { incref(); }
+
+    ByteCounter & operator=(ByteCounter const & other)
+    {
+        if (&other != this)
+        {
+            decref();
+            m_impl = other.m_impl;
+            incref();
+        }
+
+        return *this;
+    }
+
+    ByteCounter(ByteCounter && other)
+      : m_impl(other.m_impl)
+    { incref(); }
+
+    ByteCounter & operator=(ByteCounter && other)
+    {
+        if (&other != this)
+        {
+            decref();
+            m_impl = other.m_impl;
+            incref();
+        }
+
+        return *this;
+    }
+
+    ~ByteCounter() { decref(); }
+
+    void swap(ByteCounter & other)
+    {
+        std::swap(m_impl, other.m_impl);
+    }
+
+    void increase(std::size_t amount)
+    {
+        m_impl->allocated += amount;
+    }
+
+    void decrease(std::size_t amount)
+    {
+        m_impl->deallocated += amount;
+    }
+
+    std::size_t bytes() const { return m_impl->allocated - m_impl->deallocated; }
+    std::size_t allocated() const { return m_impl->allocated; }
+    std::size_t deallocated() const { return m_impl->deallocated; }
+    /* This is for debugging. */
+    std::size_t refcount() const { return m_impl->refcount; }
+
+private:
+
+    void incref() { ++m_impl->refcount; }
+
+    void decref()
+    {
+        if (nullptr == m_impl)
+        {
+            // Do nothing.
+        }
+        else if (1 == m_impl->refcount)
+        {
+            delete m_impl;
+            m_impl = nullptr;
+        }
+        else
+        {
+            --m_impl->refcount;
+        }
+    }
+
+    ByteCounterImpl * m_impl;
+
+}; /* end class ByteCounter */
+
+template <class T>
+struct MyAllocator
+{
+
+    using value_type = T;
+
+    // Just use the default constructor of ByteCounter for the data member
+    // "counter".
+    MyAllocator() = default;
+
+    static ByteCounter counter;
+
+    T * allocate(std::size_t n)
+    {
+        if (n > std::numeric_limits<std::size_t>::max() / sizeof(T))
+        {
+            throw std::bad_alloc();
+        }
+        const std::size_t bytes = n*sizeof(T);
+        T * p = static_cast<T *>(std::malloc(bytes));
+        if (p)
+        {
+            counter.increase(bytes);
+            return p;
+        }
+        else
+        {
+            throw std::bad_alloc();
+        }
+    }
+
+    void deallocate(T* p, std::size_t n) noexcept
+    {
+        std::free(p);
+
+        const std::size_t bytes = n*sizeof(T);
+        counter.decrease(bytes);
+    }
+
+    ByteCounter counter;
+
+}; /* end struct MyAllocator */
+
+template <class T, class U>
+bool operator==(const MyAllocator<T> & a, const MyAllocator<U> & b)
+{
+    return a.counter == b.counter;
+}
+
+template <class T, class U>
+bool operator!=(const MyAllocator<T> & a, const MyAllocator<U> & b)
+{
+    return !(a == b);
+}
+
+class Matrix
+{
+public:
+    Matrix();
+    ~Matrix();
+
+    Matrix(size_t, size_t, std::vector<double> const);
+    Matrix(size_t, size_t);
+
+    double& operator()(size_t, size_t) const;
+    bool operator==(const Matrix) const;
+    bool operator!=(const Matrix) const;
+
+    Matrix transpose() const;
+
+    const size_t& n_row() const;
+    const size_t& n_col() const;
+
+    friend Matrix multiply_mkl(Matrix const&, Matrix const&);
+private:
+    // Member functions
+    // Member data.
+    size_t m_col, m_row;
+    double* m_buffer;
+}; /* end class Line */
+
+class TileHelper
+{
+public:
+    Matrix padMatrix(size_t, const Matrix&);
+private:
+}; /* end class Line */
+
+Matrix TileHelper::padMatrix(size_t blocksize, const Matrix& originalMatrix) {
+    size_t row, col;
+    row = originalMatrix.n_row() + ((blocksize - (originalMatrix.n_row() % blocksize)) % blocksize);
+    col = originalMatrix.n_col() + ((blocksize - (originalMatrix.n_col() % blocksize)) % blocksize);
+    
+    Matrix result(row, col);
+    for (size_t i = 0; i < row; ++i) {
+        for (size_t j = 0; j < col; ++j) {
+            if (i >= originalMatrix.n_row() || j >= originalMatrix.n_col()) result(i, j) = 0;
+            else result(i,j) = originalMatrix(i, j);
+        }
+    }
+    return result;
+}
+
+Matrix::Matrix() {
+    this->m_row = 0;
+    this->m_col = 0;
+    this->m_buffer = NULL;
+}
+
+Matrix::~Matrix() {
+    this->m_row = 0;
+    this->m_col = 0;
+}
+
+Matrix::Matrix(size_t row, size_t col, std::vector<double> const value) {
+    this->m_row = row;
+    this->m_col = col;
+
+    int cnt = 0;
+
+    this->m_buffer = (double*)malloc(sizeof(double) * value.size());
+    memset(this->m_buffer, 0, sizeof(double) * value.size());
+
+    for (const double& v : value) {
+        this->m_buffer[cnt++] = v;
+    }
+}
+
+Matrix::Matrix(size_t row, size_t col) {
+    this->m_row = row;
+    this->m_col = col;
+
+    size_t size = row * col;
+
+    this->m_buffer = (double*)malloc(sizeof(double) * size);
+    memset(this->m_buffer, 0, sizeof(double) * size);
+}
+
+Matrix Matrix::transpose() const {
+    Matrix result(this->m_col, this->m_row);
+
+    for (size_t i = 0; i < this->m_row; ++i) {
+        for (size_t j = 0; j < this->m_col; ++j) {
+            size_t idx = this->m_col * i + j;
+            result(j, i) = this->m_buffer[idx];
+        }
+    }
+
+    return result;
+}
+
+bool Matrix::operator==(const Matrix rhs) const {
+    if (this->n_col() != rhs.n_col()) return false;
+    if (this->n_row() != rhs.n_row()) return false;
+
+    int col = this->m_col, row = this->m_row;
+
+    for (int i = 0; i < col; ++i) {
+        for (int j = 0; j < row; ++j) {
+            size_t idx = this->m_col * i + j;
+            double diff = abs(this->m_buffer[idx] - rhs(i,j));
+            if (diff > 0.0000001) return false;
+        }
+    }
+
+    return true;
+}
+
+bool Matrix::operator!=(const Matrix rhs) const {
+    return !(*this == rhs);
+}
+
+
+double& Matrix::operator()(size_t r, size_t c) const {
+    size_t idx = r * this->m_col + c;
+    return this->m_buffer[idx];
+}
+
+const size_t& Matrix::n_row() const {
+    return this->m_row;
+}
+
+const size_t& Matrix::n_col() const {
+    return this->m_col;
+}
 
 Matrix multiply_naive(const Matrix& lhs, const Matrix& rhs) {
     if (lhs.n_col() != rhs.n_row()) {
